Move tests to separate directories and run them separately
