+++ .travis.yml
@@ -16,9 +16,21 @@ jobs:
       script:
         - ./bin/scalafmt --test
         - sbt scalafixCheck docs/docusaurusCreateSite
-    - env: TEST="sbt slow/test"
+    - env: TEST="sbt 'slow/testOnly -- tests.sbt'"
       script:
-        - sbt slow/test
+        - sbt 'slow/testOnly -- tests.sbt'
+    - env: TEST="sbt 'slow/testOnly -- tests.mill'"
+      script:
+        - sbt 'slow/testOnly -- tests.mill'
+    - env: TEST="sbt 'slow/testOnly -- tests.gradle'"
+      script:
+        - sbt 'slow/testOnly -- tests.gradle'
+    - env: TEST="sbt 'slow/testOnly -- tests.maven'"
+      script:
+        - sbt 'slow/testOnly -- tests.maven'
+    - env: TEST="sbt 'slow/testOnly -- tests.feature'"
+      script:
+        - sbt 'slow/testOnly -- tests.feature'
     - env: TEST="sbt +cross/test"
       script:
         - sbt +cross/test
