+++ tests/slow/src/test/scala/tests/feature/CompletionCrossSlowSuite.scala
@@ -1,6 +1,7 @@
-package tests
+package tests.feature
 
 import scala.meta.internal.metals.{BuildInfo => V}
+import tests.BaseCompletionSlowSuite
 
 object CompletionCrossSlowSuite
     extends BaseCompletionSlowSuite("completion-cross") {
