+++ tests/slow/src/test/scala/tests/gradle/GradleSlowSuite.scala
@@ -1,4 +1,4 @@
-package tests
+package tests.gradle
 
 import scala.concurrent.Future
 import scala.meta.internal.builds.GradleDigest
@@ -8,6 +8,7 @@ import scala.meta.internal.metals.ServerCommands
 import scala.meta.internal.metals.{BuildInfo => V}
 import scala.meta.io.AbsolutePath
 import scala.meta.internal.builds.GradleBuildTool
+import tests.BaseImportSuite
 
 object GradleSlowSuite extends BaseImportSuite("gradle-import") {
